fix appBar and responsive UI
