+++ frontend/src/components/AppBar/style.css
@@ -1,0 +1,9 @@
+.app-bar {
+  position: sticky;
+  background-color: white;
+  top: 0;
+  left: 0;
+  right: 0;
+  padding-top: 10px;
+  z-index: 100;
+}
