+++ frontend/src/components/BottomBar/style.css
@@ -1,8 +1,10 @@
 .bottom-bar {
-  position: absolute;
-  bottom: 10px;
+  position: sticky;
+  background-color: white;
+  bottom: 0;
   left: 0;
   right: 0;
+  padding: 5px 0px 10px;
   border-top: 0.1em solid #bdbdbd;
 }
 
@@ -19,4 +21,12 @@
   display: flex;
   justify-content: center;
   align-items: center;
+}
+
+.bottom-bar__items a {
+  color: black;
+}
+
+.bottom-bar__items .active {
+  color: #7f57ff;
 }
